split ganim into individual components
