--- ganim.py
@@ -1,915 +1,0 @@
-"""
-ganim: Animating the history of a file using Git.
-
--------------------------------------------------
-
-This is free and unencumbered software released into the public domain.
-
-Anyone is free to copy, modify, publish, use, compile, sell, or distribute this software,
-either in source code form or as a compiled binary, for any purpose, commercial or
-non-commercial, and by any means.
-
-In jurisdictions that recognize copyright laws, the author or authors of this software
-dedicate any and all copyright interest in the software to the public domain. We make
-this dedication for the benefit of the public at large and to the detriment of our heirs
-and successors. We intend this dedication to be an overt act of relinquishment in
-perpetuity of all present and future rights to this software under copyright law.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
-INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
-PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM,
-DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-
-For more information, please refer to <http://unlicense.org/>
-"""
-
-from typing import Dict, List, NamedTuple
-
-from dataclasses import dataclass, field
-from argparse import ArgumentParser
-from asyncio.futures import Future
-from asyncio import ensure_future
-from bisect import bisect_left
-from asyncio import sleep
-from pathlib import Path
-from enum import Enum
-
-from pydriller import Repository, ModificationType  # type: ignore
-from pygments.styles import get_all_styles  # type: ignore
-from rich.syntax import Syntax
-from rich.text import Text
-
-from textual.views._window_view import WindowView
-from textual.layouts.grid import GridLayout
-from textual.scrollbar import ScrollBar
-from textual.reactive import Reactive
-from textual.geometry import clamp
-from textual._easing import EASING
-from textual.widget import Widget
-from textual.view import View
-from textual.app import App
-from textual import events
-
-
-# ganim data structures -----------------------------------------------------------------
-
-
-class Modification(NamedTuple):
-    """
-    namedtuple representing a file modification
-
-    old_path: Path | None
-        None if file was created, else same as new_path
-    new_path: Path | None
-        None if file was deleted, else same as old_path
-    type: pydriller.ModificationType
-        see pydriller docs
-    added: Dict[int, str]
-        added/modified lines in the form of (line number, content), sorted
-    deleted: Dict[int, str]
-        deleted lines in the form of (line number, content), sorted
-    """
-
-    old_path: Path | None
-    new_path: Path | None
-    type: ModificationType
-    added: Dict[int, str]
-    deleted: Dict[int, str]
-
-
-class Commit(NamedTuple):
-    """
-    namedtuple representing a git commit
-
-    author: str
-        author username
-    message: str
-        commit message
-    modifications: List[Modification]
-        list of ganim.Modification representing modifications pushed in commit
-    """
-
-    author: str
-    message: str
-    modifications: List[Modification]
-
-
-@dataclass
-class File:
-    """
-    dataclass representing a file
-
-    path: Path
-        path to file
-    current_line: int = 0
-        used to keep track of which line ContentView should resume animation of file from
-    content: List[str] = []
-        file contents as a List[str], so line contents can be easily changed using
-        indexing
-    current: bool = False
-        used by ganim.FileManager to keep track of whether this file was modified during
-        the current commit
-    deleted: bool = False
-        same as `current` argument however for deletion; file will be removed from
-        FileManager.files after update
-    """
-
-    path: Path
-    current_line: int = 0
-    content: List[str] = field(default_factory=list)
-    current: bool = False
-    deleted: bool = False
-
-
-class ModificationIterationMethod(Enum):
-    """iteration method enum"""
-
-    TOP_BOTTOM = "top_bottom"
-    NEAREST = "nearest"
-
-
-class Behaviour(NamedTuple):
-    """ganim behaviour namedtuple, set from clargs"""
-
-    # ganim args
-    targets: List[Path] = []
-    repo_root: Path = Path("")
-    easing_style: str = "in_out_cubic"
-    easing_duration: float = 0.75
-    wpm: int = 500
-    quit_once_done: int = -1
-    iter_method: ModificationIterationMethod = ModificationIterationMethod.NEAREST
-
-    # rich.Syntax args
-    theme: str = "default"
-    line_numbers: bool = True
-    indent_guides: bool = True
-    word_wrap: bool = False
-
-    # pydriller.Repository args
-    from_commit: str | None = None
-    to_commit: str | None = None
-    from_tag: str | None = None
-    to_tag: str | None = None
-    only_in_branch: str | None = None
-    only_no_merge: bool = False
-    only_authors: List[str] | None = None
-    only_commits: List[str] | None = None
-    only_releases: bool = False
-    filepath: str | None = None
-    only_file_types: List[str] | None = None
-
-
-# textual widgets -----------------------------------------------------------------------
-
-
-class FileManager(Widget):
-    """custom widget that shows and manages open files"""
-
-    spacing: str = "  "
-    files: Dict[Path, File] = {}
-
-    def render(self) -> Text:
-        file_text = Text(justify="left", overflow="crop")
-
-        # unique everseen
-        ue_files = []
-        seen = set()
-
-        for p, f in self.files.items():
-            if str(p) in seen:
-                continue
-            seen.add(str(p))
-            ue_files.append(f)
-
-        for ind, file in enumerate(ue_files[:-15:-1]):  # get last 15 files
-            style: str = ""
-
-            # underline current files
-            if file.current:
-                style = "underline"
-
-                if ind == 0:  # bold first file
-                    style += " bold"
-
-            # dim old files
-            else:
-                style = "dim"
-
-            if file.deleted:
-                style += " strike italic"
-
-            file_text.append(Text.from_markup(f"[{style}]{file.path.name}[/]"))
-            file_text.append(self.spacing)
-
-        return file_text
-
-    async def update(
-        self, mod_type: ModificationType, old_path: Path | None, new_path: Path | None
-    ) -> File:
-        """
-        updates internal file list if needed and returns ganim.File object for
-        modification by ContentView.ganimate()
-        """
-
-        if mod_type == ModificationType.ADD and new_path is not None:  # file is added
-            self.files.update(
-                {
-                    new_path: File(
-                        path=new_path,
-                        current_line=0,
-                        current=True,
-                    )
-                }
-            )
-            return self.files[new_path]
-
-        # file was deleted
-        elif mod_type == ModificationType.DELETE and old_path is not None:
-            # self.files[old_path].deleted = True
-            # return self.files[old_path]
-            self.files[old_path].deleted = True
-            return self.files[old_path]
-
-        elif old_path is not None and new_path is not None:
-            _file = self.files.pop(old_path)
-            _file.current = True
-            _file.path = new_path
-            self.files.update({new_path: _file})
-            return self.files[old_path]
-
-        else:
-            print("unreachable")
-            raise Exception("unreachable (no suitable new_path old_path criteria)")
-
-    async def advance(self) -> None:
-        """
-        called before an animation of a new commit to False the current property of any
-        file modifed previous commit, and to remove them from self.files if it was
-        deleted previous commit
-        """
-        for p, f in list(self.files.items()):
-            if f.current:
-                f.current = False
-            if f.deleted:
-                self.files.pop(p)
-
-
-class CommitInfo(Widget):
-    """custom widget that shows the current commit"""
-
-    text: str = ""
-
-    def render(self) -> Text:
-        return Text.from_markup(self.text)
-
-
-class ContentView(View):
-    """
-    custom textual view object based on textual.widgets.ScrollView
-
-    animates new changes to file objects returned by FileWidget.update() based on file
-    modifications (ganim.Modification)
-    """
-
-    def __init__(
-        self,
-        name: str | None = None,
-    ) -> None:
-        self.fluid = True
-        self.vscroll = ScrollBar(vertical=True)
-        self.hscroll = ScrollBar(vertical=False)
-        self.window = WindowView("", auto_width=False, gutter=(0, 0))
-        layout = GridLayout()
-        layout.add_column("main")
-        layout.add_row("main")
-        layout.add_areas(content="main,main")
-        super().__init__(name=name, layout=layout)
-
-    x: Reactive[float] = Reactive(0.0, repaint=False)
-    y: Reactive[float] = Reactive(0.0, repaint=False)
-
-    target_x: Reactive[float] = Reactive(0.0, repaint=False)
-    target_y: Reactive[float] = Reactive(0.0, repaint=False)
-
-    @staticmethod
-    def _clamp(value: int | float, minimum: int | float, maximum: int | float):
-        """replaces textual.geometry.clamp"""
-        if minimum > maximum:
-            maximum, minimum = minimum, maximum
-        if value < minimum:
-            return minimum
-        elif value > maximum:
-            return maximum
-        else:
-            return value
-
-    def validate_x(self, value: float) -> float:
-        return clamp(value, 0, self.max_scroll_x)
-
-    def validate_target_x(self, value: float) -> float:
-        return clamp(value, 0, self.max_scroll_x)
-
-    def validate_y(self, value: float) -> float:
-        return clamp(value, 0, self.max_scroll_y)
-
-    def validate_target_y(self, value: float) -> float:
-        return clamp(value, 0, self.max_scroll_y)
-
-    @property
-    def max_scroll_y(self) -> float:
-        return max(0, self.window.virtual_size.height)
-
-    @property
-    def max_scroll_x(self) -> float:
-        return max(0, self.window.virtual_size.width)
-
-    async def watch_x(self, new_value: float) -> None:
-        self.window.scroll_x = round(new_value)
-        self.hscroll.position = round(new_value)
-
-    async def watch_y(self, new_value: float) -> None:
-        self.window.scroll_y = round(new_value)
-        self.vscroll.position = round(new_value)
-
-    async def on_mount(self, event: events.Mount) -> None:
-        assert isinstance(self.layout, GridLayout)
-        self.layout.place(
-            content=self.window,
-        )
-        await self.layout.mount_all(self)
-
-    # FIXME: is there a better way to do this?
-    async def update(self) -> None:
-        """rebuilds file contents and updates contentview window"""
-        if self.file is not None and self.syntax is not None:
-
-            self.syntax.code = "\n".join(self.file.content).rstrip()
-            await self.window.update(self.syntax)
-
-    async def register_file(self, file: File, syntax: Syntax):
-        """sets self.file, self.syntax and updates contentview window"""
-        self.file = file
-        self.syntax = syntax
-        await self.window.update(syntax)
-
-    async def scroll_to(
-        self,
-        line: int,
-        easing: str,
-        duration: float,
-    ) -> None:
-        """scrolls to a specific line
-
-        line: int
-            line to scroll to
-        easing: str
-            easing method
-        duration: float
-            duration in seconds
-        """
-        from math import ceil
-
-        res = line - self.size.height // 2
-        self.target_y = line - self.size.height // 2
-
-        # animate if l, else
-        if abs(self.target_y - self.y) > 1:
-            self.animate("y", res, easing=easing, duration=duration)
-        else:
-            self.y = self.target_y
-
-
-# textual app
-
-
-class GAnim(App):
-    """ganim textual app"""
-
-    def __init__(
-        self, *args, behaviour: Behaviour, commits: List[Commit], **kwargs
-    ) -> None:
-        self.behaviour: Behaviour = behaviour
-        self.commits: List[Commit] = commits
-        self.ganimate_future: Future | None = None
-
-        super().__init__(*args, **kwargs)
-
-    async def on_load(self) -> None:
-        """key bindings"""
-        await self.bind("f", "view.toggle('fileview')")
-        await self.bind("c", "view.toggle('commitview')")
-        await self.bind("q", "quit")
-
-    async def on_mount(self) -> None:
-        """mount widgets"""
-        # create
-        self.contentview = ContentView(name="ContentView")
-        self.commitinfo = CommitInfo(name="CommitInfo")
-        self.filemgr = FileManager(name="FileManager")
-
-        # dock
-        await self.view.dock(self.filemgr, edge="top", size=1, name="fileview")
-        await self.view.dock(self.commitinfo, edge="bottom", size=1, name="commitview")
-        await self.view.dock(self.contentview)
-
-        # get ready
-        await self.call_later(self.start_ganimate)
-
-    async def action_quit(self) -> None:
-        if self.ganimate_future is not None:
-            self.ganimate_future.cancel()
-        return await super().action_quit()
-
-    async def on_shutdown_request(self, event: events.ShutdownRequest) -> None:
-        if self.ganimate_future is not None:
-            self.ganimate_future.cancel()
-        return await super().on_shutdown_request(event)
-
-    async def start_ganimate(self):
-        """
-        bootstraps GAnim.ganimate() call using asyncio.ensure_future()
-
-        stores returned future in self.ganimate_future
-        """
-        self.ganimate_future = ensure_future(self.ganimate())
-
-    async def ganimate(self) -> None:
-        """where the magic happens"""
-        if len(self.commits) == 0:
-            self.commitinfo.text = f"[bold red]no commits to analyse, nothing to do"
-            self.commitinfo.refresh()
-
-            if self.behaviour.quit_once_done != -1:
-                await sleep(self.behaviour.quit_once_done)
-                await self.action_quit()
-
-        spc: float = 60 / (self.behaviour.wpm * 4.7)  # seconds per character
-
-        # FIXME: is there a better way to do this?
-        async def _refresh():
-            """
-            nested function to refresh filemanager, contentview and contentinfo all at
-            once, used only during animation
-            """
-            self.commitinfo.refresh()
-            self.filemgr.refresh()
-            await self.contentview.update()
-
-        for commit in self.commits:
-            await self.filemgr.advance()
-
-            # update CommitInfo widget
-            self.commitinfo.text = f"[bold cyan]<{commit.author}>[/]: {commit.message}"
-            self.commitinfo.refresh()
-
-            # process modified files
-            for mod in commit.modifications:
-                # 0.2.0
-                # await self.commitview.animod(mod)
-
-                # update filemgr
-                file = await self.filemgr.update(
-                    mod_type=mod.type, old_path=mod.old_path, new_path=mod.new_path
-                )
-                self.filemgr.refresh()
-
-                # create syntax
-                syntax = Syntax(
-                    code="",
-                    lexer=Syntax.guess_lexer(str(file.path)),
-                    theme=self.behaviour.theme,
-                    line_numbers=self.behaviour.line_numbers,
-                    indent_guides=self.behaviour.indent_guides,
-                    word_wrap=self.behaviour.word_wrap,
-                )
-
-                # register file with contentview
-                await self.contentview.register_file(file, syntax)
-
-                # skip animation if whole file was deleted
-                if mod.type == ModificationType.DELETE and len(mod.deleted) == len(
-                    self.contentview.file.content
-                ):
-                    continue
-
-                # FIXME: is there a better way to do this?
-                last_line: int = file.current_line
-                for ln, added, deleted in moditer(
-                    mod, method=self.behaviour.iter_method, position=file.current_line
-                ):
-                    clen = len(self.contentview.file.content)
-
-                    # move to position
-                    await self.contentview.scroll_to(
-                        line=ln,
-                        easing=self.behaviour.easing_style,
-                        duration=self.behaviour.easing_duration,
-                    )
-
-                    if deleted != "":
-                        for _ in deleted:
-                            self.contentview.file.content[
-                                ln - 1
-                            ] = self.contentview.file.content[ln - 1][:-1]
-
-                            # refresh and wait
-                            await _refresh()
-                            await sleep(spc)
-
-                    if added != "":
-                        # create any needed lines
-                        if clen == 0 and ln == 1:
-                            self.contentview.file.content = [""]
-                        elif ln > clen:
-                            for _ln in range(clen - 1, ln - 1):
-                                self.contentview.file.content.append("")
-
-                        for c in added:
-                            self.contentview.file.content[ln - 1] += c
-
-                            # refresh and wait
-                            await _refresh()
-                            await sleep(spc)
-
-                    last_line = ln
-
-                file.current_line = last_line
-
-        # finish up; dim text to signal completion of animation
-        await self.filemgr.advance()
-        self.commitinfo.text = f"[dim]{self.commitinfo.text}"
-        await _refresh()
-
-        if self.behaviour.quit_once_done != -1:
-            await sleep(self.behaviour.quit_once_done)
-            await self.action_quit()
-
-
-# ganim utilities -----------------------------------------------------------------------
-
-
-def moditer(
-    mod: Modification,
-    method: ModificationIterationMethod = ModificationIterationMethod.TOP_BOTTOM,
-    position: int = 0,
-):
-    """
-    custom iterator for file modifications
-
-    mod: Modification,
-        ganim.Modification object
-    method: ModificationIterationMethod = ModificationIterationMethod.TOP_BOTTOM
-        ganim.ModificationIterationMethod object
-    position: int = 0
-        used if iter method is nearest_*, if so change value from 0 to the cursor
-        position before last modification animation
-    """
-    last = max(max(mod.added.keys(), default=-1), max(mod.deleted.keys(), default=-1))
-
-    if last == -1:
-        yield 0, "", ""
-
-    elif method == ModificationIterationMethod.TOP_BOTTOM:
-        for ln in range(1, last):
-            deleted = mod.deleted.get(ln, "")
-            added = mod.added.get(ln, "")
-            yield ln, added, deleted
-
-    else:
-        # get nearest modifed line
-        nearest = _get_nearest(
-            sorted(list(mod.added.keys()) + list(mod.deleted.keys())),
-            position,
-        )
-
-        for ln in range(nearest, last + 1):
-            yield ln, mod.added.get(ln, ""), mod.deleted.get(ln, "")
-
-        for ln in range(nearest):
-            yield ln, mod.added.get(ln, ""), mod.deleted.get(ln, "")
-
-
-def _get_nearest(l: List[int], n: int) -> int:
-    pos = bisect_left(l, n)
-    if pos == 0:
-        return l[0]
-    elif pos == len(l):
-        return l[-1]
-    else:
-        pos_left, pos_right = l[pos - 1], l[pos]
-        return pos_left if (pos_left - n) < (n - pos_right) else pos_right
-
-
-def process_args() -> Behaviour:
-    """handles clargs and sets a global Behaviour variable as bev"""
-    default = Behaviour()
-
-    parser = ArgumentParser(
-        prog="ganim", description="animating the history of a file using git"
-    )
-    # ganim arguments
-    parser.add_argument(
-        "targets",
-        help="target file paths, defaults to all files",
-        type=Path,
-        nargs="*",
-        default=default.targets,
-    )
-    parser.add_argument(
-        "--repo_root",
-        help="path to repo, defaults to cwd",
-        type=Path,
-        default=default.repo_root,
-    )
-    parser.add_argument(
-        "--easing_style",
-        help="specify textual easing style",
-        choices=[ek for ek in sorted(EASING.keys())],
-        type=str,
-        default=default.easing_style,
-    )
-    parser.add_argument(
-        "--easing_duration",
-        help=f"specify easing duration, defaults to {default.easing_duration}",
-        type=float,
-        default=default.easing_duration,
-    )
-    parser.add_argument(
-        "--wpm",
-        help=f"specify words per minute, defaults to {default.wpm}",
-        type=int,
-        default=default.wpm,
-    )
-    parser.add_argument(
-        "--quit_once_done",
-        help=(
-            "quits n seconds after animation finishes, "
-            f"defaults to {default.quit_once_done} (dont quit)"
-        ),
-        type=int,
-        default=-1,
-    )
-    parser.add_argument(
-        "--iter_method",
-        help=f"specify line iteration method, defaults to {default.iter_method.value}",
-        choices=[im.value for im in ModificationIterationMethod],
-        type=str,
-        default=default.iter_method.value,
-    )
-
-    # syntax arguments
-    syntax_args = parser.add_argument_group(
-        "code syntax args",
-        description="see https://rich.readthedocs.io/en/stable/syntax.html",
-    )
-    syntax_args.add_argument(
-        "--theme",
-        help="specifies a pygments theme to display file contents in, see https://pygments.org/styles/",
-        choices=[theme for theme in get_all_styles()],
-        type=str,
-        default=default.theme,
-    )
-    syntax_args.add_argument(
-        "--line_numbers",
-        help=f"enable indent guides, defaults to {default.line_numbers}",
-        choices=[True, False],
-        type=bool,
-        default=default.line_numbers,
-    )
-    syntax_args.add_argument(
-        "--indent_guides",
-        help=f"enable indent guides, defaults to {default.indent_guides}",
-        choices=[True, False],
-        type=bool,
-        default=default.indent_guides,
-    )
-    syntax_args.add_argument(
-        "--word_wrap",
-        help=f"enable word wrapping, defaults to {default.word_wrap}",
-        choices=[True, False],
-        type=bool,
-        default=default.word_wrap,
-    )
-
-    # commit range arguments
-    commit_range_args = parser.add_argument_group(
-        "commit range arguments",
-        description=(
-            "see https://"
-            "pydriller.readthedocs.io/en/latest/repository.html"
-            "#selecting-the-commit-range"
-        ),
-    )
-    commit_range_args.add_argument(
-        "--from_commit", help="starting commit", type=str, default=default.from_commit
-    )
-    commit_range_args.add_argument(
-        "--to_commit", help="ending commit", type=str, default=default.to_commit
-    )
-    commit_range_args.add_argument(
-        "--from_tag",
-        help="starting the analysis from specified tag",
-        type=str,
-        default=default.from_tag,
-    )
-    commit_range_args.add_argument(
-        "--to_tag",
-        help="ending the analysis from specified tag",
-        type=str,
-        default=default.to_tag,
-    )
-
-    # commit filter arguments
-    commit_filter_args = parser.add_argument_group(
-        "commit_filter_args",
-        description=(
-            "see https://"
-            "pydriller.readthedocs.io/en/latest/repository.html#filtering-commits"
-        ),
-    )
-    commit_filter_args.add_argument(
-        "--only_in_branch",
-        help="only analyses commits that belong to this branch",
-        type=str,
-        default=default.only_in_branch,
-    )
-    commit_filter_args.add_argument(
-        "--only_no_merge",
-        help=(
-            "only analyses commits that are not merge commits, "
-            f"defaults to {default.only_no_merge}"
-        ),
-        choices=[True, False],
-        type=bool,
-        default=default.only_no_merge,
-    )
-    commit_filter_args.add_argument(
-        "--only_authors",
-        help="only analyses commits that are made by these authors' usernames",
-        nargs="+",
-        type=str,
-        default=default.only_authors,
-    )
-    commit_filter_args.add_argument(
-        "--only_commits",
-        help="only these commits will be analyzed",
-        nargs="+",
-        type=str,
-        default=default.only_authors,
-    )
-    commit_filter_args.add_argument(
-        "--only_releases",
-        help=(
-            "only commits that are tagged will be analyzed, "
-            f"defaults to {default.only_releases}"
-        ),
-        choices=[True, False],
-        type=bool,
-        default=default.only_releases,
-    )
-    commit_filter_args.add_argument(
-        "--filepath",
-        help="only commits that modified this file will be analyzed",
-        type=str,
-        default=default.filepath,
-    )
-    commit_filter_args.add_argument(
-        "--only_file_types",
-        help="only show histories of certain file types, e.g. '.py'",
-        nargs="+",
-        type=str,
-        default=default.only_file_types,
-    )
-
-    args = parser.parse_args()
-
-    # file type standardisation
-    only_file_types: List[str] | None = None
-    if args.only_file_types is not None:
-        only_file_types = []
-        for file_type in args.only_file_types:
-            if file_type.startswith("."):
-                only_file_types.append(file_type)
-            else:
-                only_file_types.append(f".{file_type}")
-
-    return Behaviour(
-        targets=args.targets,
-        repo_root=args.repo_root,
-        easing_style=args.easing_style,
-        easing_duration=args.easing_duration,
-        wpm=args.wpm,
-        quit_once_done=args.quit_once_done,
-        iter_method=ModificationIterationMethod(args.iter_method),
-        theme=args.theme,
-        line_numbers=args.line_numbers,
-        indent_guides=args.indent_guides,
-        word_wrap=args.word_wrap,
-        from_commit=args.from_commit,
-        to_commit=args.to_commit,
-        from_tag=args.from_tag,
-        to_tag=args.to_tag,
-        only_in_branch=args.only_in_branch,
-        only_no_merge=args.only_no_merge,
-        only_authors=args.only_authors,
-        only_commits=args.only_commits,
-        only_releases=args.only_releases,
-        filepath=args.filepath,
-        only_file_types=only_file_types,
-    )
-
-
-def process_repo(behaviour: Behaviour) -> List[Commit]:
-    """processes the repository for files to animate, returns List[Commit]"""
-
-    commits: List[Commit] = []
-
-    for commit in Repository(
-        path_to_repo=str(behaviour.repo_root),
-        from_commit=behaviour.from_commit,
-        to_commit=behaviour.to_commit,
-        from_tag=behaviour.from_tag,
-        to_tag=behaviour.to_tag,
-        only_in_branch=behaviour.only_in_branch,
-        only_no_merge=behaviour.only_no_merge,
-        only_authors=behaviour.only_authors,
-        only_commits=behaviour.only_commits,
-        only_releases=behaviour.only_releases,
-        filepath=behaviour.filepath,
-        only_modifications_with_file_types=behaviour.only_file_types,
-    ).traverse_commits():
-        modifications: List[Modification] = []
-
-        for file in commit.modified_files:
-            # skip file if targets were specified and file is not a target
-            if len(behaviour.targets) > 0 and file.filename not in [
-                str(p) for p in behaviour.targets
-            ]:
-                continue
-
-            # skip file if file types were specified and file type was not specified
-            elif (
-                behaviour.only_file_types is not None
-                and Path(file.filename).suffix not in behaviour.only_file_types
-            ):
-                continue
-
-            diff = file.diff_parsed
-            old_path = Path(file.old_path) if file.old_path is not None else None
-            new_path = Path(file.new_path) if file.new_path is not None else None
-            added: Dict[int, str] = {}
-            deleted: Dict[int, str] = {}
-
-            for line, content in diff["added"]:
-                added.update({line: content})
-
-            for line, content in diff["deleted"]:
-                deleted.update({line: content})
-
-            modifications.append(
-                Modification(
-                    old_path=old_path,
-                    new_path=new_path,
-                    type=file.change_type,
-                    added=added,
-                    deleted=deleted,
-                )
-            )
-        if len(modifications) > 0:
-            commits.append(
-                Commit(
-                    author=commit.author.name,
-                    message=commit.msg,
-                    modifications=modifications,
-                )
-            )
-
-    return commits
-
-
-# ganim entry point ---------------------------------------------------------------------
-
-
-def main() -> None:
-    """ganim entry point, returns Behaviour"""
-
-    behaviour: Behaviour = process_args()
-    commits: List[Commit] = process_repo(behaviour)
-
-    try:
-        import uvloop
-
-    except ImportError:
-        pass
-
-    else:
-        uvloop.install
-
-    GAnim.run(behaviour=behaviour, commits=commits)
-
-
-if __name__ == "__main__":
-    main()
