+++ src/ganim/__init__.py
@@ -1,0 +1,9 @@
+# ganim: Animating a file's history using Git.
+#
+# This file is part of ganim, which is dedicated to the public domain.
+# See <https://unlicense.org/> for more information or the UNLICENSE file included with
+# the ganim source code.
+
+from . import structures, widgets
+from .app import GAnim
+from .cli import run
